requirements.txt
streamlit
groq
python-dotenv
sounddevice
numpy
scipy
speechrecognition
